% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{calculateDE.sg}
\alias{calculateDE.sg}
\title{calculateDE.sg}
\usage{
calculateDE.sg(data, info, cpcol, sf, gene, onlyoneSample)
}
\arguments{
\item{data}{data}

\item{info}{info}

\item{cpcol}{cpcol}

\item{sf}{sf}

\item{gene}{gene}

\item{onlyoneSample}{onlyoneSample}
}
\value{
This return the paralelled calculated DE result.
}
\description{
This is an internal function for doDE
}
\examples{

}
