% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sampleFile.R
\name{createSampFile}
\alias{createSampFile}
\title{Create Sample Files}
\usage{
createSampFile(fileName1, fileName2, sampleNames, write.to.file = FALSE,
  output.name = "sampleFiles")
}
\arguments{
\item{fileName1}{For single-end reads, a vector of all fastq files. For paired end, a vector of all fastq that are first in pair}

\item{fileName2}{Vector of filepaths for all fastq for reads that are second in their paired reads. If empty, will treat files in \code{fileName1} as single-end}

\item{sampleNames}{Vector of names for samples}

\item{write.to.file}{Logical indicating if the table should be written to a file, which is used for the quantification function}

\item{output.name}{User-specified filename for the output file}
}
\value{
A table with either 2 (single-end) or 3 (paired-end) columns
}
\description{
This function creates a file telling the package which files to obtain sequence data from.
Depending on the user's input, the output will be catered to single or paired read experiments
}
\note{
Ensure that the number of reads matches up with the number of sample names given. Also, due to
constraints of \code{runKallisto()}, the sample file created can only be saved to where the fastq files are.
It is also necessary to keep all fastq files in the same directory
}
\author{
Matthew Myint
}
