% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{Donbregression}
\alias{Donbregression}
\title{Donbregression}
\usage{
Donbregression(data, info, cpcol, reftype, sf, gene, onlyoneSample)
}
\arguments{
\item{data}{data}

\item{info}{info}

\item{cpcol}{cpcol}

\item{reftype}{reftype}

\item{sf}{sf}

\item{gene}{gene}

\item{onlyoneSample}{onlyoneSample}
}
\value{
This return the nb calculated p value and fold change
}
\description{
This is an internal function for doDE
}
\examples{

}
