% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezSeurat.R
\name{ezSeurat}
\alias{ezSeurat}
\title{Launch shinyApp}
\usage{
ezSeurat()
}
\value{
Launches the shiny app
}
\description{
Use this function to run the shinyApp
}
\examples{
ezSeurat()
}
\author{
Matthew Myint
}
\keyword{shiny}
