% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezSingleCell.R
\name{ezSingleCell}
\alias{ezSingleCell}
\title{Launch shinyApp}
\usage{
ezSingleCell()
}
\value{
Launches the shiny app
}
\description{
Use this function to run the shinyApp
}
\examples{
ezSingleCell()
}
\author{
Matthew Myint
}
\keyword{shiny}
