% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackstraw_pval.R
\name{JackStraw_pval}
\alias{JackStraw_pval}
\title{Get PC scores}
\usage{
JackStraw_pval(object, PCs = 1:20, score.thresh = 1e-05)
}
\arguments{
\item{object}{The seurat object}

\item{PCs}{PCs to select from}

\item{score.thresh}{emperical p-value cut-off to apply}
}
\value{
A dataframe of PCs and their p-values
}
\description{
Calculate scores for PC based on jackstraw results
}
\examples{
pc_pval <- JackStraw_pval(object = sObj)
}
\author{
Matthew Myint, Jinmiao Chen
}
