% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Index_kallisto.R
\name{kallisto_index}
\alias{kallisto_index}
\title{Build kallisto index}
\usage{
kallisto_index(index.name, fasta.file, kallisto.path = "kallisto",
  specify.kmer.size = NULL, make.unique = FALSE, output.dir = getwd(),
  dry.run = FALSE)
}
\arguments{
\item{index.name}{Specify a filename for the output, filetype .idx will be appended to it}

\item{fasta.file}{Path to the fasta file that will be used to build the index, can be gzipped}

\item{kallisto.path}{This argument is to be used if kallisto is not found in the \code{PATH} variable}

\item{specify.kmer.size}{The length of k-mer to be used in the index. Must be an odd number between 3 and 31}

\item{make.unique}{Logical argument on whether repeated targets should be given unique names, defaults to false}

\item{output.dir}{Path to directory to save output index file to}

\item{dry.run}{If true, will return the command to be passed to kallisto}
}
\value{
If \code{dry.run = TRUE}, prints the command to be passed to kallisto, otherwise creates an index file at target directory
}
\description{
This function creates a kallisto index, for later use to do read quantification.
}
\author{
Matthew Myint
}
