% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE.R
\name{rbindlist}
\alias{rbindlist}
\title{rbindlist}
\usage{
rbindlist(pardoresult, info, cpcol)
}
\arguments{
\item{pardoresult}{pardoresult}

\item{info}{info}

\item{cpcol}{cpcol}
}
\value{
return a list that includes all DE result iteratively
}
\description{
This is an internal function for doDE that combine result from paralell computation
}
\examples{

}
