% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackstraw_pval.R
\name{signif_PCs}
\alias{signif_PCs}
\title{PC filtering}
\usage{
signif_PCs(score.df, p.val = 0.05)
}
\arguments{
\item{score.df}{A dataframe of PCs and their scores}

\item{p.val}{p-value cut-off for significant PCs}
}
\value{
A numeric vector of significant PCs
}
\description{
Choose significant PCs based on their p-value for use in TSNE
}
\examples{
score.df <- data.frame(PC = paste0("PC", 1:10), Score = runif(10, 0, 0.2))
pc_pval <- signif_PCs(object = sObj)
}
\author{
Matthew Myint, Jinmiao Chen
}
