% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{violin_plot}
\alias{violin_plot}
\title{Create interactive violin plot from Seurat object}
\usage{
violin_plot(sObj, type, interactive = TRUE, event.data = NULL)
}
\arguments{
\item{sObj}{The seurat object}

\item{type}{The parameter to be visualised. Currently applies to metadata (nGene, nUmi, percent.mito)}

\item{interactive}{If \code{TRUE}, output will be plotly object, else a ggplot object.}
}
\value{
A plotly object
}
\description{
Can hover over points to see the cell ID, its current identity, original identity, and the value of the parameter being plotted.
}
\examples{

}
\author{
Matthew Myint
}
